+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 1.1.2, 4/2/2022
+
+### Added
+
+Forward source IP address to target
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 1.1.1, 11/25/2021
 
 ### Added